update websocket example
